Build and submit transaction
