+++ rust/main/Cargo.toml
@@ -1,27 +1,27 @@
 [workspace]
 members = [
-  "agents/relayer",
-  "agents/scraper",
-  "agents/validator",
-  "applications/hyperlane-application",
-  "applications/hyperlane-operation-verifier",
-  "applications/hyperlane-warp-route",
-  "chains/hyperlane-cosmos",
-  "chains/hyperlane-cosmos-native",
-  "chains/hyperlane-ethereum",
-  "chains/hyperlane-fuel",
-  "chains/hyperlane-sealevel",
-  "ethers-prometheus",
-  "hyperlane-base",
-  "hyperlane-core",
-  "hyperlane-metric",
-  "hyperlane-test",
-  "submitter",
-  "utils/abigen",
-  "utils/backtrace-oneline",
-  "utils/crypto",
-  "utils/hex",
-  "utils/run-locally",
+    "agents/relayer",
+    "agents/scraper",
+    "agents/validator",
+    "applications/hyperlane-application",
+    "applications/hyperlane-operation-verifier",
+    "applications/hyperlane-warp-route",
+    "chains/hyperlane-cosmos",
+    "chains/hyperlane-cosmos-native",
+    "chains/hyperlane-ethereum",
+    "chains/hyperlane-fuel",
+    "chains/hyperlane-sealevel",
+    "ethers-prometheus",
+    "hyperlane-base",
+    "hyperlane-core",
+    "hyperlane-metric",
+    "hyperlane-test",
+    "submitter",
+    "utils/abigen",
+    "utils/backtrace-oneline",
+    "utils/crypto",
+    "utils/hex",
+    "utils/run-locally",
 ]
 resolver = "2"
 
@@ -61,10 +61,10 @@ config = "0.13.3"
 console-subscriber = "0.2.0"
 convert_case = "0.6"
 cosmrs = { version = "0.21.0", default-features = false, features = [
-  "cosmwasm",
-  "rpc",
-  "tokio",
-  "grpc",
+    "cosmwasm",
+    "rpc",
+    "tokio",
+    "grpc",
 ] }
 cosmwasm-std = "*"
 crunchy = "0.2"
@@ -126,15 +126,15 @@ rlp = "=0.5.2"
 rocksdb = "0.21.0"
 rstest = "0.25.0"
 sea-orm = { version = "1.1.10", features = [
-  "sqlx-postgres",
-  "runtime-tokio-native-tls",
-  "with-bigdecimal",
-  "with-time",
-  "macros",
+    "sqlx-postgres",
+    "runtime-tokio-native-tls",
+    "with-bigdecimal",
+    "with-time",
+    "macros",
 ] }
 sea-orm-migration = { version = "1.1.10", features = [
-  "sqlx-postgres",
-  "runtime-tokio-native-tls",
+    "sqlx-postgres",
+    "runtime-tokio-native-tls",
 ] }
 semver = "1.0"
 serde = { version = "1.0", features = ["derive"] }
@@ -219,27 +219,27 @@ overflow-checks = true
 [workspace.dependencies.ethers]
 features = []
 git = "https://github.com/hyperlane-xyz/ethers-rs"
-tag = "2025-05-02"
+tag = "2025-05-06"
 
 [workspace.dependencies.ethers-contract]
 features = ["legacy"]
 git = "https://github.com/hyperlane-xyz/ethers-rs"
-tag = "2025-05-02"
+tag = "2025-05-06"
 
 [workspace.dependencies.ethers-core]
 features = []
 git = "https://github.com/hyperlane-xyz/ethers-rs"
-tag = "2025-05-02"
+tag = "2025-05-06"
 
 [workspace.dependencies.ethers-providers]
 features = []
 git = "https://github.com/hyperlane-xyz/ethers-rs"
-tag = "2025-05-02"
+tag = "2025-05-06"
 
 [workspace.dependencies.ethers-signers]
 features = ["aws"]
 git = "https://github.com/hyperlane-xyz/ethers-rs"
-tag = "2025-05-02"
+tag = "2025-05-06"
 
 [patch.crates-io.curve25519-dalek]
 branch = "v3.2.2-relax-zeroize"
