+++ rust/main/submitter/src/chain_tx_adapter/chains/ethereum/transaction.rs
@@ -1,0 +1,5 @@
+pub use factory::TransactionFactory;
+pub use precursor::Precursor;
+
+mod factory;
+mod precursor;
