+++ rust/main/submitter/src/chain_tx_adapter/chains/factory.rs
@@ -4,12 +4,11 @@
 use std::sync::Arc;
 
 use eyre::Result;
+
 use hyperlane_base::{
     settings::{ChainConf, ChainConnectionConf, RawChainConf},
     CoreMetrics,
 };
-use hyperlane_core::{ContractLocator, HyperlaneDomain, HyperlaneDomainProtocol, H256};
-use hyperlane_ethereum::{EvmProviderForSubmitter, SubmitterProviderBuilder};
 
 use crate::chain_tx_adapter::{
     chains::{cosmos::CosmosTxAdapter, ethereum::EthereumTxAdapter, sealevel::SealevelTxAdapter},
@@ -24,8 +23,6 @@ impl ChainTxAdapterFactory {
         raw_conf: &RawChainConf,
         metrics: &CoreMetrics,
     ) -> Result<Arc<dyn AdaptsChain>> {
-        use HyperlaneDomainProtocol::*;
-
         let adapter: Arc<dyn AdaptsChain> = match conf.connection.clone() {
             ChainConnectionConf::Ethereum(connection_conf) => Arc::new(
                 EthereumTxAdapter::new(conf.clone(), connection_conf, raw_conf.clone(), metrics)
