+++ rust/main/submitter/src/payload_dispatcher/test_utils.rs
@@ -3,11 +3,10 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use eyre::Result;
-use hyperlane_base::db::{DbResult, HyperlaneRocksDB, DB};
+use hyperlane_base::db::{HyperlaneRocksDB, DB};
 use hyperlane_core::identifiers::UniqueIdentifier;
 use hyperlane_core::KnownHyperlaneDomain;
 use tokio::sync::Mutex;
-use uuid::Uuid;
 
 use super::*;
 use crate::chain_tx_adapter::*;
